refactor: 移除 RecipeDraft 相關元件以簡化結構，並更新引用路徑以改善檔案組織。
